# Improve SeqDecoder performance by removing overhead