Hi, I was fiddling around to see if there was some way of making Circe faster and found that the current SeqDecoder has a non-trivial amount of overhead.

I believe this was mostly due to cursor object creation and destruction. But I can't be sure as those don't appear in the profiling flame graph that I used.

To compare, I ran `sbt 'benchmark/jmh:run -f1 io.circe.benchmark.NumberParsingBenchmark.decodeMany*'` before and after the optimization and there is a trend towards making arrays of primitives faster:

```
# before
[info] Benchmark                                      Mode  Cnt      Score     Error  Units
[info] NumberParsingBenchmark.decodeManyBigDecimals  thrpt    5   1520.022 ±  87.924  ops/s
[info] NumberParsingBenchmark.decodeManyBigInts      thrpt    5   1773.855 ±  12.936  ops/s
[info] NumberParsingBenchmark.decodeManyDoubles      thrpt    5  10004.552 ±  28.427  ops/s
[info] NumberParsingBenchmark.decodeManyLongs        thrpt    5  12490.381 ± 133.195  ops/s

# after
[info] Benchmark                                      Mode  Cnt      Score    Error  Units
[info] NumberParsingBenchmark.decodeManyBigDecimals  thrpt    5   1478.384 ± 62.771  ops/s
[info] NumberParsingBenchmark.decodeManyBigInts      thrpt    5   1799.242 ±  8.659  ops/s
[info] NumberParsingBenchmark.decodeManyDoubles      thrpt    5  10451.115 ± 19.050  ops/s
[info] NumberParsingBenchmark.decodeManyLongs        thrpt    5  13264.556 ± 55.887  ops/s
```

Just to confirm, I made a JMH test in the style of the previous benchmark class with booleans:
```scala
@State(Scope.Thread)
@BenchmarkMode(Array(Mode.Throughput))
@OutputTimeUnit(TimeUnit.SECONDS)
class ArraysBenchmark {
  val count = 1000

  val inputBooleans = "[" + List.fill(count)(true).mkString(", ") + "]"

  @Benchmark
  def decodeManyBoolean(): Either[circe.Error, Array[Boolean]] = jawn.decode[Array[Boolean]](inputBooleans)
}
```

and the effects are quite dramatic:

```
# before
[info] Benchmark                           Mode  Cnt      Score     Error  Units
[info] ArraysBenchmark.decodeManyBoolean  thrpt    5  38027.942 ± 273.341  ops/s

# after
[info] Benchmark                           Mode  Cnt      Score     Error  Units
[info] ArraysBenchmark.decodeManyBoolean  thrpt    5  56613.658 ± 293.886  ops/s
```